Move scripting wrapper
