+++ apps/extension/src/BackgroundScript/misc/forumPageLinks.ts
@@ -1,4 +1,4 @@
-import scripting from '@helpers/chrome/scripting';
+import scripting from '@/utils/scripting';
 
 const getForumPageLinksFunc = () => {
   const unreadRows = document.querySelectorAll(
